requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.32
pytest==8.3.2
requests==2.32.5
streamlit==1.48.1
python-dotenv==1.1.1
pydantic==2.11.7